dom/camera/test/test_bug1037322.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1037322</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id="viewfinder" width="200" height="200" autoplay></video>
<img src="#" alt="This image is going to load" id="testimage"/>
<script class="testbody" type="text/javascript;version=1.7">

var suite = new CameraTestSuite();

suite.test('bug-1037322', function() {
  var cameraManager = navigator.mozCameras;
  var whichCamera = cameraManager.getListOfCameras()[0];

  var postConfig = {
    mode: 'picture',
    recorderProfile: 'low',
    previewSize: {
      width: 320,
      height: 240
    }
  };

  function resolveGetCamera(p) {
    suite.camera = p.camera;

    // Check the default configuration
    var cfg = p.configuration;
    ok(cfg.mode === "unspecified", "Initial mode = " + cfg.mode);
    ok(cfg.previewSize.width === 0 && cfg.previewSize.height === 0,
       "Initial preview size = " + cfg.previewSize.width + "x" + cfg.previewSize.height);
    ok(cfg.recorderProfile === "default",
       "Initial recorder profile = '" + cfg.recorderProfile + "'");
  }

  function configure(p) {
    // Apply our specific configuration
    return suite.camera.setConfiguration(postConfig);
  }

  function resolveConfigure(cfg) {
    // Check our specific configuration
    ok(cfg.mode === postConfig.mode, "Configured mode = " + cfg.mode);
    ok(cfg.previewSize.width === postConfig.previewSize.width &&
       cfg.previewSize.height === postConfig.previewSize.height,
       "Configured preview size = " + cfg.previewSize.width + "x" + cfg.previewSize.height);
    ok(cfg.recorderProfile === postConfig.recorderProfile,
       "Configured recorder profile = '" + cfg.recorderProfile + "'");
  }

  return cameraManager.getCamera(whichCamera, {mode: 'unspecified'})
    .then(resolveGetCamera, suite.rejectGetCamera)
    .then(configure)
    .then(resolveConfigure, suite.rejectConfigure);
});

suite.setup()
  .then(suite.run);

</script>
</body>

</html>
